frontend: Allow a list of groups/version/resource in API factory methods
We want to support different K8s API versions, and sometimes all they
differ in is the version and/or the group used in the API call.

In order to make this change seamless, this patch adapts the API
factory methods, to support multiple group/version/resource arguments,
and use the next one in the list if there's been an 404 error
accessing them.
i.e. if the first API group/version/resource attempted fails with 404,
it may be because the API version is deprecated, so it will try the
next one in the list. This should have no effect in other cases of 404
errors, since it always ends up sending the lsat 404 error found if
none of the calls is successful.