frontend: Ensure that the wildroute is considered in the last place
When routes are used with the switch component from react-router, they
are used sequentially to see which one matches the URL, and if a
wildcard route is found (as used for the 404) before other routes, it
will always match, essentially blocking other routes from being
considered.

This patch prevents that by excluding the wildcard route from the
default routes, and using it as the last route considered in the router
switch.