frontend: Give precedence to dynamic routes when matching them
We have default routes and dynamic ones (set in redux, by plugins), and
so far we've been considering the default ones first. This meant that
routes couldn't be overridden by plugins and this makes things less
flexible for them.

This patch considers default routes last, so it opens a lot of
flexibility for plugins to replace a route completely.