frontend: Allow to apply unknown resource configurations
Headlamp was using its resource classes in order to apply the
configurations desired by the user (in the Create/Edit dialog).
However, if the user used a configuration of a Custom Resource
Definition, it would fail as the kind would not be known to Headlamp.

This patch fixes that by not requiring the use of resource classes
for the mentioned uses.