app: Allow to run the app with or without starting the server
Using the app in dev mode meant not starting the server from the app,
but sometimes we do need to do that in order to verify the exact
behavior we will get when the app is run in production.

So this patch renames the "npm run serve-dev command" to "npm run dev"
(which runs the server), and adds a new "npm run dev-only-app" which
does not run the server.