frontend: Do not mount UI before i18n is loaded
When the UI was loaded, there were lots of messages about missing
translation keys. This was due to the translation function being used
before all the namespaces were loaded.

So we now verify if the i18n is loaded, before mounting the actual UI.
While it's not shown, we show a spinner.