frontend: Reset plugins state before initializing them
We had an issue related to re-initializing the plugins when they've
already be initialized, this ended up re-registering the same plugin
functionality and thus would result in getting duplicated sections
and potentially other issues.

While re-initialization of plugins should be avoided altogether, the
logic for initializing them should make sure the plugins and their
states are reset before initialization which is what this patch does.