frontend: Centralize k8s listing requests
As the user navigates the UI, there may be several calls to k8s
server for listing and getting resources. This means that a quick
navigation may retrieve new objects that were already in memory.
This not only implies more IO but may also lead to more memory usage
by Headlamp on the client side.

Thus, these changes centralize the listing requests (only these ones
for now, but we should add them all in the future), using a ref count
to check when requests need to be stopped/renewed, which allows for a
more efficient memory and network usage.