backend: Fix oidc redirect url token type
In an oidc auth flow the id token is used for authentication and that is
what should be send in the redirect URI of the oidc callback so that
id_token is used to make request to the kubernetes cluster.
Access token should not be used because they are part of oauth 2.0 flow
and are used for authorization purposes.