plugins/headlamp-plugin: Do not ignore types when packing
We want to ignore the presence of generated types (types folder) in
git, but we want them to be shipped in the module. However, npm does
check .gitignore if a .npmignore file is missing, which meant that no
types were being packed/shipped.

This commit adds a .npmignore file that doesn't ignore the types,
effectively making npm pack the mentioned types.