frontend: Do not use objects for storing plugins functions
Throughout the Redux code, we were using objects to store a map of
name -> something, but keys' order is not kept in objects, so we could
not predict which actions were being called after others.

This patch replaces that use with a Map subclass, so we can still use
it very similarly to what we used to, but now we guarantee that the
values can be used in insertion order.

As a side-effect, the serialization of a Map is more predictable than
that of an object (the Map is always {} by default), so we have a more
predictable use of the data within Redux (this comment has to do with
the fact that we shouldn't use unserializable data in Redux, but that
should be for cases where one's thinking of storing/persisting Redux's
values, which is not the case for this particular UI data).