backend: Add no-cache header for plugins when running in desktop
When running in the desktop, there is no big advantage to allowing the
plugins to be cached, and we have hit issues where the plugins were
being cached while the user is attempting to reload them (for
development). So this patch prevents that by adding a no-cache header
to the /plugins/ request responses.