plugins/headlamp-plugin: Set include to from the top directory's src
The include path was relative to headlamp-plugin in node_modules, so
it didn't work when just extending the plugins-tsconfig.json file.
This change set the path to be the plugin's own top directory + src
files, so it allows plugins' tsconfig to simply extend the mentioned
shipped tsconfig instead of having to override the "includes".