frontend: Add better a11y error reporting
Issues are logged last in the console. To make them easier to read.

If any issues remain, an alert() is done to warn developers when
headlamp is compiled in development mode. This alert() can be
disabled with the REACT_APP_SKIP_A11Y=true environment variable,
which is necessary when beginning development on new components.

Hopefully these two measures will help prevent more issues escaping
review, since the console log was quite noisy and can be forgotten.

Started a section in the documentation on a11y tools.