app: Kill Headlamp server process synchronously
When the app is quitting, if we try to kill the Headlamp server
process asynchronously, there's a race condition which may end up in
the electron process being quit before the server kill instruction is
actually done. The result was that the Headlamp server would be left
running in Windows most of the times.