app: Ask if a running Headlamp process should be killed
Until recently, there was a good chance that Headlamp would leave its
backend running after the app was quit (due to a bug).

Even though that bug should be fixed, we want to make sure that any
users who've run an older version of Headlamp, do not run into issues
with the newer versions due to that dangling process, so this patch
checks if such a process is running and asks the user whether to kill
it.