frontend: Only alert about a11y issues once per session
If there are multiple a11y issues reported by axe, then the UI
becomes unusable with the alert notifications that need to be
acknowledged.
This may lead developers to just run all the time with the skipped
a11y alerts.

So this patch tries to prevent that by allowing only a one-time alert
per session (i.e. if the user refreshes, they should see it again).