frontend: Do not duplicate children in dynamic sidebar items
When items were set up in the sidebar from a plugin, the object they
were using was the one set up in redux. This meant that any children
of that sidebar entry would be appended to the object's "subList" in
redux and hence that list would grow with multiple equal entries
every time the sidebar was rendered.

This patch prevents that by deep cloning the object before using it.