frontend: Make the namespaces filter values occupy just one line
When multiple namespaces were selected, the input could show several
lines and that made the header for the contents very large vertically,
which didn't look good.

This patch changes how the namespaces are shown, so that only up to
a certain number of chars are visible for the namespace name and after
that it shows how many other namespaces are selected.