frontend: Add a new ResourceTable component
This component is the same as the SimpleTable but with a couple of
improvements to make it easier to use with kubernetes resources.
i.e. it allows to use certain types as columns, like "age" and "name"
which will translate into the commonly used age and name columns
respectively.

It can also be used with a resource class directly, allowing the user
to skip getting the list of resources + error before listing them.