frontend,app: Check for updates only if configured to do so
This patch loads environment variables from a .env file and ships the
file in the app as well, in order to allow different builds with minimal
changes to Headlamp's core.
Those options, when needed, are passed to the renderer process by IPC
through the recenly added appConfig message.

So these changes also only check for updates if the checkForUpdates
option in the app/packages.json is defined.