frontend: Show navigation tabs + disable sidebar button on width=sm
When the sidebar is the narrow version (when width=sm), then we should
always show the navigation tabs, otherwise the user has no way to
navigation.

The sidebar button (Headlamp icon by default) was also active even
when in the narrow sidebar version mode, meaning it would toggle the
navigation tabs, so this patch also prevents that by disabling the
button in that mode.