app: Ensure default menus are still translated on plugin overrides
Even though we have translations for the default menu items, because
of the way the mechanism for overriding menus is done it means that
we'd always keep the English menu even for default items when the
menu was overridden (even without changing any of the items) by a
plugin.

This patch prevents that by updating the menus' labels with the
translated ones, using a new "id" for each default menu in order to
match the labels.