app: Make certain menus dependent on whether the plugins are loaded
Some menus shouldn't be shown until we're sure plugins may have
changed them. This is particularly important for menus such as help,
which may carry brand-related actions, like opening the app's website,
etc.

To ensure this, this patch adds the ability to mark menus as to be
loaded only after plugins, which will give the opportunity for plugins
to change them if needed.