frontend: Add default filterFunction to ResourceTable
The ResourceTable always has the filter header and is supposed to offer
the search/filter capabilities by default but, due to a mistake, this
feature was not included when we added it.

This patch fixes that by adding a default filterFunction to the
mentioned component (callers can still override it).