frontend: Add useGet hook to KubeObject class
Like the useList hook, this is a more convenient way of getting the
information about a particular k8s resource than having to create
states for using with useApiGet.

Another motivation for this is that we want to be able to test by
mocking the "get details" function, like we are already doing with
useList in some storybook items.