headlamp-plugin: Fix broken npm audit fix handling, warn instead
If there is a package without a security fix, then this will not
work. Even if headlamp-plugin updates package.json versions.

The dependency resolution for sub-dependencies of packages does
not seem to work the same way. Also "overrides" field is ignored.

For now run npm audit fix, but if it fails only warn about it.
People will have to fix their dependencies manually, or wait
for a better upgrade.