frontend: Ensure cluster actions are only processed when they change
Due to how the useSelector from redux works, we were triggering
changes to the ActionsNotifier whenever anything in the redux state
was changing, so this resulted in continuously processing the cluster
actions even when those didn't change. This had side-effects like
always moving moving back to the "startUrl" when changing routes
while deleting a resource.

This patch adds the evaluation function to the mentioned selector, so
it will only trigger when actually the actions change.

We should still move the URL processing of the actions to somewhere
where it can be better controlled, or to remove it completely and
allow actions to define callbacks for each action state instead
(cancelled/errored), meaning they'd call the route changes themselves
when desired.