frontend: Do not hide filter header search when its contents are empty
There's something going on with the filter header so when it's hidden
by removing the search text completely with the backspace it will
make the UI scroll down.

Because of this but also due to the fact that we shouldn't really
hide the view on backspace (only on escape but this behavior was not
consistent among the namespace and the search entries either), we're
stopping the behavior of hiding the filter header altogether.