frontend: Optimize KubeObjectClass.useGet so it updates iff needed
This function was setting its state every time new data was returned
by the API function, but that data may be the same as before, which
would still trigger a change in any component using it. Thus, this
patch makes it so that the state only changes if the resourceVersion
changes.