frontend: Fix setting pod logs in LogViewer
The logs were getting set whenever they changed, and in order for the
logs not to be repeated the LogViewer clears the logs before setting
them, but if this is done repeatedly, the UI may flash and the
experience is not so smooth.

This is fixed by directly controlling the terminal and continuously
write only the logs that need to be written in it.