frontend: Introduce the PreviousRouteProvider context
This context (and useHasPreviousRoute hook) allows to get the location
from which we ended up in the current page.
This is needed in order to be able to access the mentioned previous
path and the main need right now is that we can use this info to
decide whether to go back to the previous route or not.