frontend: Add reflectInURL property to the SimpleTable
These changes allow the SimpleTable to reflect its page and
rowsPerPage properties in the URL, so views can quickly set that up
and make sure users can go back to previous views that reflect the
table pages they were reading.

This patch also adds a couple of stories to cover this use-case.