frontend: Show/hide filter depending on the search + who triggered it
When the user moves to a different route, the search filter is reset
(to an empty string), but this goes through redux, so when the new
route is rendered, the search filter is still the old one. Eventually,
the redux state will be updated, but at this point we already have the
filter being shown (when it should actually be shown). OTOH, if we
always hide the filter when the search (and namespace) are empty, it
will also hide when the user deletes the whole string, and that's not
desireable (because they may want to keep writing and at that point the
filter has een hidden and input is lost).

To solve this, we keep track of whether the filter has been shown by
the user, in which case we don't hide it even when the search is empty.