frontend: Don't replace the location when updating URL params
If we update the location + search when the URL params change, the
react-router's history won't work as we were expecting when the hash
history is in use (Electron). Apparently it won't take into account
just the hash.

To avoid this, and keep the same functionality (of reflecting the
desired states in the URL) both on the hash and the browser router,
we should use an object with just the search key when calling
history.replace.