frontend: Prevent failure when describing CronJob's schedule
The schedule is expected to be in the crontab's format but users have
reported that they are using the components with variants of CronJob
which have a single `*`, thus making the description fail.
Overall, recovering from an unknown cron format seems like a good
defensive measure, so this patch adds that.