frontend: Allow terminal to reconnect on Enter if failed to connect
When the terminal cannot be connected and runs through all the
shells available, there is no way to have it retry all the shells
(in order to recover from a temporary issue).

This patch allows to do that by asking the user to press enter if
they want to reconnect.