frontend: Use a new AuthBasedButton for permission-based actions
We have some actions and buttons whose visibility depend on whether
the user has permissions to perform that action.

So this patch adds a new component AuthBasedButton which wraps a
component that is only displayed depending on the permissions' check.

It also replaces the permissions logic in the Edit/Delete buttons in
with it.