frontend: Refresh config if there is an update on it
We now add the facility to refresh the config if there
is an update on it. This is done by fetching the config
after a certain interval of time. This also handles the
case where if the cluster somehow got deleted and the user
goes to that page we show a feedback to the user for that.

Co-authored-by: ashu8912 <aghildiyal@microsoft.com>