frontend: Don't send all routes without parameters to the chooser
Due to a previous fix where we wanted send all routes without params
to the chooser, in order to prevent sending the user to routes that
needed parameters but didn't have them, we were inadvertently sending
all users to the chooser route.

This check is no needed because we do an earlier check for whether the
cluster is necessary, and thus this patch removes it.