frontend: Fix doc viewer breaking
With a recent CRA update node related modules are not shipped
with CRA anymore. This broke the doc viewer as it was using the buffer
module. We already had a patch earlier that added the polyfill
but one problem with polyfilling it in a place like index.tsx is that
when building the frontend for production, the polyfill is not included.
The correct place to polyfill is to include it at the place where it is
being used.