frontend: Use ISO format for dates tests
Node has changes how they format the dates when using "toLocaleString",
introducing a special character (U+202f) instead of a space.
Our first instinct was to lower the Node version and wait for a fix,
but according to [1], Node doesn't seem to be reverting it soon and
mentions one shouldn't use localized strings for testing...

So, in order to avoid this, this patch makes the dates show up in
ISO format.

[1] https://github.com/nodejs/node/issues/46123