frontend: Fix the logic for checking if the pod is portforwarding
We were not considering the usecase where a portforward is part of
localStorage and not part of the backend portforward list api.
In that case that portforward is actually in a stopped state, and we
should consider the storage portforwards too when checking if the
current pod is portforwarding or not. This patch also fixes some
variable names to align with what they do.