frontend: Allow the EditButton to check for its permissions
The EditButton was always being shown as a view button because the
AuthVisible button never got to mount the AuthVisible component.
These changes make the EditButton not read-only by default, meaning
it'll rely on the AuthVisible at first to check for permissions, and
only after a first check will it either show the edit or view
buttons.