frontend: Remove actionFilterFunc from SimpleTable
This property was mistakenly added in a PR but it shouldn't be needed
as we can just use the already supported filterFunc property.

In order to make the Even filtering work, this patch also uses a
combination of the default filtering and the warning check for
maintaining the intended event filtering functionality.