frontend: Skip a11y alert by default
The purpose of the a11y alert we have when running Headlamp in
default mode is so we find regressions in a11y. However, we do have
some a11y issues as of now and the popup is just shown every time at
the beginning of running Headlamp, which prompts everyone to just
dismiss it automatically hence defeating its purpose of showing up
right when a11y regressions happen. All this means a11y issues keep
growing, which render the alert just an annoyance.

So this patch disables the alert for now until we fix the a11y
issues.