frontend: Add attach to pod
We add a new button to the pod details page that allows the user to
attach to the pod's container(by default the first available one),
using the terminal component. We also update the terminal component
to accept a new prop called isAttach which allows us to differentiate
b/w a exec and a attach session.