frontend: Improve how different sidebars are used
This patch removes the "special treatment" of the out-of-cluster
sidebar that's used in notifications and settings, and reimplements
the logic to show it in a way that is can be extended with different
sidebars.

It also allows adding sidebar items to sidebars different from the
default (in-cluster) one. I.e. plugins are now able to add sidebar
items to the notifications/settings sidebar.