backend: Add helper to generate clientset from rest config
We want to be able to use the clientset to interact with the cluster
from the backend. This adds a helper to generate a clientset from the
rest config. This patch also takes care of handling if there is a token
coming from the frontend, then modify the config to use the token.