ci: add multi-platform Docker images
Refactor the Docker image to support cross-compiling and update
the CI pipeline to publish arm64 images in addition to the amd64
images already being published.

There's also some general improvements here, e.g. the Go builds
copy `go.{mod,sum}` and run `go mod download`, cache volumes,
use of `COPY --link`, removal of unnecessary base images.

This removes the symlink for `/headlamp/server`, which has been
around for almost a year for compatibility. (Keeping it will add
a dependency on QEMU to run the `cd` / `ln` commands, as it's the
only command run in the final image, which is for the _target_
arch, not the _build_ arch.)

CI also uses GitHub Actions cache now - free speed-up!

The builds are also done with `pull: true` to ensure that the
latest Go / Node / Alpine patch versions get used. (This replaces
the `DOCKER_IMAGE_BASE` which was actually overriding things with
an _older_ version, but thankfully not in CI.)

Additionally, the images publish provenance metadata.

Signed-off-by: Milas Bowman <devnull@milas.dev>