backend: Fix externalproxy endpoint, also make it not cache requests
So we do two things in this patch, first we don't use the reverseProxy
approach because that approach is not required here, we call it the
/externalproxy endpoint but all we do is make request to the external
server from our server so that we can bypass the CORS issue that we
encounter from making request directly from frontend. Also using
reverseProxy was leading us to not transfer every request parameters
correctly so for some proxy request like
"https://artifacthub.io/api/v1/packages/helm/headlamp/headlamp" the
request would immediately fail (the missing request parameter that are
not taken care of with reverseProxy is still unknown to me).