frontend: Clone the clusters conf to prevent direct state changes
The auth chooser was getting an error when checking for a cluster's
auth because of how the cluster's state is set. It sets the state
directly and then calls dispatch, but since the object being modified
is the same that redux holds, redux detected this and it threw an
exception.

To prevent this case, the useClustersConf hook now returns a copy of
the state instead.