frontend: Prevent multiple auth test requests at login
Due to how the test-auth logic in the AuthChooser was designed, it
led to many requests to the API server, especially if the response
was not valid.

This is due to the clusters dependency in the related useEffect hook
being constantly changed. Even though this is worth an investigation
on its own, making sure these requests only happen when needed is
needed, so this patch ensures only one test auth request happens
unless the clusters are actually changed.