frontend backend: Prevent unauthorized access to helm endpoints
This prevents some backend endpoints from being called unless
there is a token passed in via the X-HEADLAMP_BACKEND-TOKEN header.
The app generates the token and then sets HEADLAMP_BACKEND_TOKEN
environment variable which the backend uses.

This also makes available a getHeadlampAPIHeaders() function
which can be used by the frontend and plugins to add headers to
API endpoints that the backend serves.

App creates token and sends it to the backend and frontend. This
prevents unauthorized connections to the helm chart endpoints.

The threat was that any connection from a Browser or localhost
process could call helm and dynamic cluster creation endpoints.