frontend: Go back to the cluster selection on error in KubeConfigLoader
If a cluster is selected to be added but it fails due to be set up in
the backend, then the user would be stuck in a dialog with an endless
spinner and an error message.
This patch goes back to the cluster selection on such errors, meaning
the user has a choice to proceed from there.