frontend: Allow setting up sidebar as null when no sidebar is needed
This fixed a regression in the behavior of the sidebar, as it was
previously possible to declare a sidebar for a route as being null and
it would not be visible until the changes to have multiple sidebars.

This patch also renames the route URL in the sidebar plugin example
to say "no sidebar" instead of "no cluster".