frontend: Simplify how the mobile topbar is set up
The "theme.breakpoints.down('xs')" in MUI v4 checks if the screen
width is under 600px (same as checking for sm), but in MUI v5 it checks
if it's under 0 (since xs does mean 0).

So this means we can drop the isMobile check as it should be the same
as the isSmall check. This patch does that and simplifies how the top
bar's logos/menus are set up accordingly.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>