frontend: make headlamp support stateless calls
Headlamp now supports stateless calls to the backend. User can create
new cluter using the plugin. This feature is only available if this
create a cluster using kubeconfig. The values of kubeconfig are stored
in session storage with the corresponsding sessionID that creates the
cluster.

When user make any calls to frontend it gets the values from the session
storage and send it as headers. Then backend creates the proxy on fly on
every request making it stateless.

Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>