Use a phony backend token when run from Makefile
We use the make file to run Headlamp in dev mode, and since the
backend token was not set up in that case, it was impossible to use
this mode for testing any backend features, i.e. running
npm run dev-only-app would not work for testing adding clusters, use
Helm, etc.