Make dynamic clusters completely stateless
Once user creates a dynamic cluster, their kubeconfig is stored in
session storage with the associated sessionID. To make it completely
stateless backend does not store it, instead they are store in redux
store and retrieved from there.

backend created proxy on the fly and send data to the frontend from k8s
api server.

Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>
Co-authored-by: yolossn <sannagaraj@microsoft.com>
Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>