frontend common/Label: Do not end-align HoverInfoLabel
The HoverInfoLabel was changed to have its contents aligned to the
end (right, in LTR languages) as the date label used should be aligned
as such and uses this component internally. However, many other values
are represented with this label, so we cannot always align it.

This patch fixes that by making this component have an inline-flex
display, allowing for the labels+icons to be vertically aligned to the
center (reason for which it had a display:flex rule) while respecting
the text-align rule set on a wrapping element. i.e. the date labels get
aligned to the right in most tables as their cell has that text-align
rule, but other uses of the mentioned component don't get affected.