frontend crd: Add new Resource list view
We now have two different list views related to Custom Resource
Definitions: the actual CRD list, and a new Custom Resource list.
The latter will show the list of objects related to the Custom
Resources (not the Definitions).

The previous CRD list view is now a mix between definitions and
resource names, allowing the user to quickly choose what they need
to see.
Cross-links between Definition and Resource are also in place for
both of the details views.