frontend: Allow applying any kind of resource
When applying a resource, we tried to get a match for a known/default
resource (CronJob, Pod, etc.) but this may result in trying to apply
a resource using the wrong API when we have a CustomResource that has
a name matching any exiting default resource's.

This patch always tries to get the API details from the server for
the API version of the data we need to apply, thus ensuring we always
get the right API.

By removing the no longer needed import of ResourceClasses in apiProxy,
this let an issue related to circular dependencies be visible which is
fixed in this commit by importing the Crd in plugin/index's pluginLib
only after the K8s module.