dynamic cluster: fix page reload
Websockets connection was getting closed as new proxy was being setup on
the fly, instead we are storing the proxy for 10minutes. If there is no
activity on the proxy we delete the proxy from the cache making it
stateless.

Fixes page reload for redux when 2 states were being setup for
kubeconfig and stateless clusters

Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>