make dynamic clusters stateless
Once user creates a dynamic cluster using the dynamic cluster plugin,
their kubeconfig is stored in session storage. This kubeconfig is
associated with user sessionID so it won't be seen in other sessions. To
make it stateless the cluster information is stored inside redux store
and data is fetched from there. Once user makes a request to the
cluster a proxy is created and stored inside cache. We have to store the
proxy in the cache to avoid closing of sockets. Every data is stored in
redux and retrieved from there.

If there is activity in the stateless cluster then the proxy is
automatically deleted from the cache to avoid garbage collection

Co-authored-by: Ashu Ghildiyal <ashu.ghildiyal@microsoft.com>
Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>