frontend useCluster: Initialize cluster to its current name
We were initializing the cluster to null and only setting its actual
name in an effect hook, which resulted in the useCluster hook firing
twice whenever used (first for null, second for the cluster name).

This resulted in the side-effect of every Resource.useList performing
two requests till the cluster hook "stabilized".

This patch initializes the cluster name in that hook to the current
one, so it will not fire twice for the same effective state.