frontend: Remove useless use of useClustersConf from useCluster
I guess the original idea of having the useClustersConf trigger the
check for useCluster was that maybe we have new configuration for any
current cluster, but the truth is that the useCluster hook would
anyway not fire because it depends on having a cluster name different
from the one in the current URL, and that's not influenced by the
useClustersConf. I.e. useClustersConf was triggering unnecessary checks
within useCluster.