frontend: Refactor makeCustomResourceClass to use the right name
The singular name used for creating the Custom Resource classes with
makeCustomResourceClass was always "crd", but this is not the correct
resource name and this results in issues when fetching the permissions
for such objects.

This commit refactors the mentioned method, allowing it to receive an
object with the needed arguments.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>