plugins: dynamic clusters example plugin update
Previously it only had a button with was taking hard coded values and
setting those values. Now backend sends information if the cluster has
dynamic clusters enabled or not. If it is enabled user are able to store
the kubeconfig statlessly.

It adds input options for setting values accordingly by the information
received from backend about isDynamicClusterEnabled.

Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>