frontend apiProxy: Add limitting to the streamResults function
If we call a function like "list" with the "limit" request param, that
will work in the first call that the listing performs internally, but
then it connects through a websocket (to stream any changes to that
list), and from that point the limit is no longer respected. This means
that a listing with a limit of X can get way more items as it remains
streaming.

This patch fixes that by doing a client side limitting on the results
coming from the streaming, guaranteeing that only the most recent X
items are returned. If no limit has been set, then it behaves as
previously (no limits are applied at all).

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>