frontend: Fix infinite loop when setting notifications
When setting notifications, it could lead to an infinite loop because
it was adding notifications from new kubernetes events whenever events
or notifications were changed. The reason was that even though the
Notifications component (App/Notifications/index.tsx) was setting
notifications with the new events, the notifications being set in Redux
+ storage could be a different list (as they are sorted by date +
limited to a number).

This patch prevents that by not updating the notifications in Redux if
they are the same that we already have, thus preventing the updates of
components that are checking the notifications and updating them as
well.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>