app: frontend: plugins: Add runCommand for shell commands in Electron app
So that some local shell commands can be run from inside a plugin.
This communicates the process interaction with the Electron app render
process via ipc.

The interface exposed is a subset of the node spawn() which is what is
used on the app side. It only implements the parts which are already
used by the plugin we have in mind.

The interface supports streaming stdout and stderr, and getting exit
codes.

Note, this supports running multiple commands at the same time. The
complexities of this are hidden to the user of runCommand.

It limits the commands that can be run to only minikube and az so
far. If there is a need for more this can be extended later. This is
intended as a way to limit potential misuse.

The app-menu example plugin has been extended to add a minikube menu
which displays minikube status.

Signed-off-by: René Dudfield <renedudfield@microsoft.com>