frontend: Disable auto reconnection of logs
Adds a new enhancement where pods logs are not auto reconnected. This
used to happen when pod's container was in retry state, which made logs
to reconnect and it kept on flashing as if the page reloaded until the
pod was in a Ready state.

This adds functionality if the pod is not Ready, it shows a button to
reconnect the logs.

Fixes: #1480
Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>