frontend: Add the cluster parameter to KubeObject's CRUD options
Due to how the cluster actions are implemented, the actual functions,
like update/delete, are only called at the end of the grace period.
Since that is also when the cluster name would be gathered, then it
could happen that the user changed clusters, and thus this could lead
to updating/deleting a resource in a cluster that is different from
the intended one (if a similar resource exists).

This patch solves this by adding the cluster name to which a resource
belongs as an inner member of that resource instance. This cluster name
is then used for the mentioned operations, hence guaranteeing that the
operations always happen for the same cluster that the item belongs to.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>