frontend: Prevent creating/applying resources to undesired clusters
When using the creation dialog to create a resource, due to the fact
that the creation function only checked the current cluster after the
grace period ends, it may happen that the user changed to a different
cluster and thus the change would be applied to a different cluster
from what was desired.

This patch fixes that by getting the cluster right from the moment the
user clicks Apply, so even if clusters are changed, the resources are
applied on the intended cluster.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>