plugins: Fix copying to docker containers
We were copying plugins but we were not installing dependecies which
resulted some example plugins to not get copied.

headlamp-events plugin was also using local dev dependency which
resulted in failure of installing it. This fixes it.

Adds test to check plugins content match that of docker images built

Since we are installing node_modules, it required significant space in
the disk for tests. So added a step to remove unnecessary files which
saved around 10GB of space.

Fixes: #1695
Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>