backend: format backend logs to use zlog
Earlier we were not using zlog everywhere due to which some logs were
not correctly parsed by log platforms. They were all parsed as errors.
This fixes that and adds correct format of all logs in headlamp. They
are now shown as json which can be parsed by any log platform.

This also adds a wrapper function for zlog which increases code
readability and cleanliness.

Fixes: #1645
Signed-off-by: Kautilya Tripathi <ktripathi@microsoft.com>